Fix a bug where some abilities are not being recognized properly as affecting mastery.
This bugfix affects detection of Holy Prism, Light of the Martyr, Holy Shock, Light's Hammer and Tyr's Deliverance, which were not being tracked properly and were not included in your Mastery Effectiveness at all. While for most players this shouldn't affect the resulting total Mastery Effectiveness by much, you will probably notice about a 2% decrease in average Mastery Effectiveness.

The reason this issue was introduced is that WCL reports different spell IDs for their casts then it does for the resulting healing effect. Luckily we had enough heals working properly to still get effective results, but this was still a major oversight and I apologize for that.

I will continue testing to ensure the accuracy of the Mastery Effectiveness Calculator.