Update license to AGPL
The goal of the license is to make it possible for other parties to use the tool within their own environment and along the way help improve it. This requires them to share the source code including that of the rest of their product so that if that may be helpful there's no excuse possible towards hiding that, which can then in turn be used to improve *this* version of the tool and possibly others.

The old license had a similar goal but technically couldn't be considered *open source* due to it prohibiting commercial usage.