Refactor of SoulShardEvents
Also deleted the targetID and targetInstance from the custom events. So far it's been unused and it would cause a bit of a confusion (since for sources that come from energize events, I would want to set undefined to them). If I would need the target in the future, I'll reintroduce it.