Add a `completeness` property to spec config that indicates how complete a spec's implementation is
Set all specs to *needs more work* by default, will do a review of a spec analysis once this is increased to see if an increase is correct. The goal of this is to make it more obvious to users how complete a spec implementation is to make it clear that some specs still need more work and aren't reflective of the tool.
This indicator might become more obvious in the future.