[Frost Mage] Thermal Void and Winters Chill Fixes
Adjusted Thermal Void to exclude the time and cast for Icy Veins if it does not complete before the fight ends as it skews the average in a negative way.

Also adjusted Winters Chill to account for "Double Ice Lance" Hits within the Winter's Chill Window (which was causing this number to go negative or potentially hide a missed window).... This also counts the number of times the player managed to get 2 Ice Lance's in the window, but i havent decided how i want to use that data yet ... so for now it counts it and doesnt do anything further with it. Ill do something with that data at some point.