[Core/FrostMage] Cancelled Casts
Module for showing the % of casts that were cancelled. Tooltip also shows the breakdown of spells cast, completed, and cancelled. Also filters out spells with a cast time, but the cast time was made instant by some proc or other event (Like Blizzard - Instant if Frozen Orb is up) .... but if the dev includes a pure instant cast spell in the list, then it will likely screw with the results because Instant Cast Spells dont have a "begincast", but spells that are sometimes instant do.