[Frost Mage] Minor Blizzard Suggestion
This adds a suggestion if you are on a fight that is primarily or partially AoE and you never cast Blizzard. This is meant to be a start to evaluating talents and spell cast choices based on the current fight ... For example, telling the player that they took a talent that gave them no benefit because its a Single Target fight, etc, etc.

I also plan on doing more with the Blizzard Module in the future, but getting this working opens up some doors to do further evaluation on the other modules, so I would like to get this in first, so i can do the others. At some point ill circle back to this and add more analysis as far as Blizzard Usage and Damage.