[Frost Mage] Misc Additions/Fixes
Added a damage counter to show how much extra damage the player got from Double Ice Lance. Also modified the cast time on Ebonbolt to include the Cast Time of the spell. That way the Cast Efficiency can be a little more accurate when it is determining how many casts were possible.