Fixed Stagger tests.
All but 1 failure was due to the Fabricator not being wired up. That 1
had its value changed due to the new way of tracking stagger. The old
one was, I believe, incorrect -- it estimated amount of stagger in pool
based on stagger ticks remaining, but the stagger ticks in the fixture
don't match what a real stagger tick would look like so the result was
off. The new way of doing it is simply to directly record damage going
into and leaving the stagger pool. Thus, the unrealistic stagger ticks
in the fixture are tracked exactly despite the fact that in a real fight
stagger would be ticking for more damage.