Fixed all (I think?) of the points raised in review.
Also went through and pre-emptively edited the IronskinBrew normalizer
to match the style as I understand it.

The one point that I did not edit for was the use of a getter for
`staggerPool`. Perhaps it is implied that crossing a Module boundary
means that you don't modify properties, but :shrug: If I get push-back
on this I'm willing to change it.