Bugfix on KegSmash module
tldr - KS is an AoE ability, so sometimes we'd count the hits multiple
times -- but we only want to know how many Stave Off procs
(pseudo-multistrike, for those not familiar) we got. After some testing,
it looks like the most reliable way get that is to only count damage
events to the primary target of the cast event.