[Fire Mage] Add Rune of Power & Adjust Cast Efficiency
Rune of Power was made previously and put in the shared folder for Mage, but was never added to the Fire CombatLogParser.

Also went through Abilities and added suggestions and recommended efficiency to some talents and abilities that didnt have it previously. Also removed cooldowns from some utility abilities to remove the Cast Efficiency % bar from them.