[Core] Fix Death Tracker
Death Tracker was not counting Soulstone as a ressurect and there are no events to indicate that the player used their soulstone. So instead, I made it so that if you are healed, take damage, start casting something, or cast something while you are marked as dead, then it will count as a ressurect.

Fixes #1207