[Core] Added Patch Compat Version
Added a Patch Compatibility Option to the CONFIG to allow the developer to indicate what patch the analyzer is certified/verified for. This will help locate outdated analyzers and let the users know whether patch changes are included or not.

Also adjusted the warning display order so feedbackWarning is higher up than the needs work warning (since only the first one will actually get displayed)

Fixes #884