Update ColdHeartEfficiency.js
Moved all the constants to the correct places and updated the & into &&.

Changed buffColdHeart to 20 to account for your character starting with 20 stacks at the start of the fight.

Added two new trackers to give better feedback of usage. One tracks when the cast was made too early and the other when the cast was made too late.