[Core] Stats, Talents, Gear Layour Changes
Moves the Stats on Pull, Talents, and Gear into a Character Panel alongside Checklist and Suggestions. This also reduces scrolling by removing Stats and Talents from the side Panel on the left side.

Additionally, I played around with the talents and stats layout.

I left the Gear Layout mostly the same and just added a heading to it, im sure more can be done to make this look better, perhaps to mirror the character sheet layout in game, but without item names or the realm name to get the character picture, this is probably good for now.

Fixes #1260
Fixes #211