[Elemental] Only warn about elemental focus for >1 Flames Shock casts
Instead of warning when a single flame shock was cast without elemental
focus, we suggest no more than 1 flame shocks are cast without it.
Reason for the change is elemental shaman's opener, as it is possible
to cast

prepot/fire elemental -> precast EB -> FS -> LB and on

leading to unbuffed initial (short) flame shock.