Update DarkTransformationAndWounds.js
Changed if statement to correctly account for target having a null property for wounds. Also changed the === true and false stuff. Might want to change the if statement altogether, to give true if any enemy has wounds on them. The tracker will start counting for quick target switching, eventhough the main target still has wounds. Not sure how to deal with that scenario, mostly Vari mythic that it applies to.