General cleanup of the Brew Generation checklist item.
This introduces a new `performanceMethod`: first. This is useful when
one of your checklist entries aggregates the others already in a
complicated way (like my brew CDR code). If you're screwing up on the
other items, the first item will be red and the overall will thus be
red. Showing the other items gives more information as to *why* it is
red (e.g. low KS casts? low BoB casts? not casting BoB?)