[Prepatch] [Frost Mage] Glacial Spike Changes
The rotation for Glacial Spike was slightly changed.

Now it is a DPS increase (something like 2-3%) to hold Glacial Spike until you have a Brain Freeze Proc to cast with it as opposed to getting as many Glacial Spike's out as possible.

Additionally, because of this, if the player is spec'd into Ebonbolt, they should hold Ebonbolt and purposely not cast it just incase they have Glacial Spike and need a proc in a pinch. Because of this, it is acceptable (if they player has good luck) for them to never cast Ebonbolt in the fight.