Healthstone Suggestion Rework
I got a complaint from the mage discord that the wording on the Healthstone Efficiency Suggestion was misleading.

The way it reads is that you should have your healthstone on cooldown for 60% of the fight, meaning you should use it early in the fight to ensure its on cooldown for over 60% of the fight.

This is likely just an example of the default cast efficiency suggestion text not sounding right when talking about spells and abilities that can only be used once. As opposed to something like Ebonbolt that should be on cooldown for 90% of the fight.

So, I created a custom suggestion for Healthstones, left it in the efficiency tab, and used the possible casts/actual casts as the trigger for the suggestion. Also changed the wording to make a little more sense and be easier to understand.

I imagine this is going to cause a conflict with the BFA Branch. Not sure if i should submit a second PR for the bfa branch? or if it will get taken care of another way?