[Arcane Mage] Mana Chart
Im not 100% sure if this is accurate or not. I pretty much just took the healing Mana chart and adjusted it to accomodate damage, but that involved copying and editing a lot of stuff into DamageDone and I am not sure if I missed anything.

Im in no rush to get this in, so I would appreciate any feedback on what I might have missed here to make this more accurate.