[Arcane Mage] Mana Chart
This adds a mana chart for Arcane. I decided against using the Mana Usage Chart as i didnt feel like the DPS Throughput was all that helpful, so i went with the Mana Level Chart instead as part of playing Arcane well is ending the fight with as little mana as possible (since you should burn your mana at the end of the fight regardless of cooldown state.