Adds basic checklist while removing inaccurate Raging Blow warning
Adds a basic checklist with the basic cooldown, downtime, and preparation sections, as well as a 'Use Rampage' section that checks whether generators are being cast when Rampage should be cast instead (>90 rage), as per the Icy Veins guide.

This also removes the Raging Blow efficiency requirement, as it can be observed from either guides or logs that Raging Blow is a low-priority spell and typically has a low (20-40%) efficiency.