[Core] Replace item() with statistic() (Part 1)
First off, i know these dont look like your example commit, but I noticed the new stat boxes had ItemStatistic so i used that instead and just duplicated the BoringSpellValueText to make an item version of it and used that instead.

Also, some of the items that were using item() were Legion Legendaries, so i just deleted those.

This isnt all of them, but it is a good start.

Towards #2939