[Core] Fix for Cancelled Casts in Timeline
There was a bug with the timeline where it would show a spell as cancelled if the player cast an ability that is castable while casting (Such as Shimmer, Fire Blast, Ice Floes, or Combustion). These spells do not interrupt spell casting and should not trigger marking an ability as cancelled.

Maybe there is a better way to resolve this, but i saw the CASTS_THAT_ARE_NOT_CASTS thing and just decided to replicate that since it was the easiest way to fix it.