Handling CP gains from Sinister Strike procs
There is a bug I noticed with WoW combat logs where the CPs gained from SS procs do not appear in an energize event.
The handler here will look for damage events from that SS proc and use the processInvisibleEnergize from ResourceTracker to account for the the CP gains.
As far as I've been able to tell, wasted CP from a SS that procs are what gets shown in the energize event, so I believe the ResourceTracker should handle this gain accurately