Some cleanup to help with building the checklist
The BTE, Dispatch, and Opportunity modules still had their suggestion thresholds using the 'number' style. I missed changing this when changing those suggestions to percentages.
The RTB Efficiency module was rearranged a little bit, so that the actual suggestion thresholds were part of the suggestion array. This makes it much easier to include in the checklist.