Add Furious Gaze documentation
https://www.warcraftlogs.com/reports/TGzmk4bXDZJndpj7#fight=6&type=auras&source=5&ability=273232&start=2497903&end=2523828&view=events
The log sometimes applies a new buff before the last one was finished.
This causes buffs to be counted multiple times so we only handle events
when we know the previous event succesfully finished

https://www.warcraftlogs.com/reports/fNL9mkXcMp1GdnQ3#fight=27&type=auras&source=176&ability=273232&start=3811599&end=3841408
When Furious Gaze is active, triggering the buff again causes the timer
to extend. Because of this we only get 1 applyBuff and 1 removeBuff
event, which would throw off the average and missed Furous Gaze
calculations