Fix eslint
This doesn't mean it's ok to use PropTypes.any, I'm only using it because I have to fix a dozen files that I didn't work in right now so don't have the context to properly fix everything within a reasonable amount of time. If you can improve these types that would be amazing.