[Core] Cast Efficiency Fix
Cast Efficiency was showing > 100% for the Time on Cooldown for a lot of spells.
This was because when calculating the Time on CD, it was essentially doing (SpellCooldown * SpellCasts / FightDuration).

So the remaining cooldown of the spell beyond the end of the fight was counting towards the time on cooldown.