Tweak BW cooldown efficiency
The spell is so short CD with all the reductions in 8.3 that it isn't uncommon for WoWA to tell you that you've used it too few times even if only a GCD is in between it coming off CD and it being cast. This tries to account for that by simply lowering the cast efficiency. It will still trigger if people aren't casting it enough, but won't trigger if there are simply a GCD between it coming off CD and it being cast.