[Mage] From the Ashes, Arcane Orb, & General Cleanup
- Added the From the Ashes Talent for Fire (Towards #3657 )
- Added Covenant Abilities to the Spellbook for all 3 specs
- Removed the GCD for Icy Veins and Arcane Power
- Adjusted Arcane Orb since it is acceptable for ST now (Fixes #3948 )
- Added types to all the variables for Frost and Fire

I also started undoing some of what i did with constants and have gone the route of using a constants file for the whole class instead of individual spec constants. Additionally, if a constant is only used in one module, I removed it from the centralized constants file and put it back into the relevant module.