[Core] Filtered Active Time/Downtime
Open to suggestions, this is working at the moment but seems to be skewed by some wonky things going on with Haste in Prepatch (Im assuming some haste values are using pre-squish haste amounts).

This adds the ability to get the active time/downtime for a specfied time frame. So for example, i could get the player's uptime while Combustion was active.

It has some other uses that I plan on looking into (Maybe forced downtime on bosses?)