Keep stdin open in dev service
Some piece of the developer server's stack assumes it is running in
a TTY and treats a closed STDIN as signal to terminate the server. I saw
multiple references to react-scripts having this problem on
stackoverflow but I didn't track down the specific layer in our case.
Either way, this was causing a newly cloned repo to fail to launch and
immediately exit the docker service via e.g. `docker-compose up dev`.
This commit directs docker-compose to open STDIN to the process running
for the `dev` service, keeping the server alive indefinitely so it can
respond to requests.