Remove null default for antiFillerSpam prop
which was being set *after* the superclass's constructor was called,
probably in part because this class is not Typescript.

As indicated in previous discussion, it's probably best to stop jumping
through hoops to get this additional data into SpellbookAbility child objects
and just provide it as a parallel type created within guardiandruid, but
that is a bigger change and I am still learning the architecture of this
project, so I thought it worthwhile to stabilize this feature in the
meantime.