fix handling of spell cooldowns in APL checker
previously, it relied on the `expectedDuration`---which is incorrect
when there is a large swing in haste (such as BL expiring). this
switches it to rely on `updatespellusable` events.

however, those events may be out of order due to the way they're
created, so this also requires adding a sort prior to functional
analysis to put them in the correct order