[WIP][Arcane Mage] Update to 9.1.5
This is a work in progress as I work on getting Arcane Mage updated for patch 9.1.5 (Doing 9.1.5 before i look at 9.2).

Additionally, a lot of what is in Arcane Mage is extremely messy and complicated to follow, so I am also working on refactoring/rewriting a lot of the spec so it is easier to maintain (it is essentially impossible for me to update it to 9.1.5 unless i clean this up).