Fix problem with the map objects.
The second object got keys that was `[object Object]` because I forgot to do `.id` on it. Not sure why typescript caught that to be perfectly honest. Changed the maps to return tuples via `as const`, and then used type parameter on fromEntries. This seems to cause error if first tuple value is not number or string.