Support splitting of ticks happening the same millisecond.
Any damage pulse can only affect any target once, if the same target is hit multiple times, it is two different events happening, somehow, at the same time. We just check for duplicate hits on the same targets and if that is the case, we pretend the new damage event happens a ms later to handle them as two separate events.