add TalentCastEfficiencyRequirement component
add CombatLogParserContext to make CombatLogParser
available via hooks for functional components.

add TalentCastEfficiencyRequirement component
that will only render if the selected combatant
has the given talent.

TalentCastEfficiencyRequirement will also fetch
the cast efficiency of the given talent so that
we don't need to set up an AbilityRequirement in
each spec checklist.

update Vengeance checklist to use it as an
example.