[Fire Mage] DF Updates Round 1
Initial Round of DF Updates. This is not all the updates needed, and some of them may be slightly inaccurate, but will be ironed out over time.

Additionally, I am aware that my suggestions, statistics, and checklist is all giving bad info right now. I didnt bother updating them because I will be replacing them with the Guide, likely in the next pass.