[Frost Mage] Fixed Winters Chill
Rewrote Winters Chill to resolve an issue where precasts and shatters were not properly being counted. This seemed to be mostly an issue if you were standing at max range and the spells took longer to get to the target than anticipated.

Also fixed an issue in the "next" function in Event History where it was grabbing more events than intended .... essentially ignoring the count variable.