[Talents] Import talentId (id) from Raidbots, and add helpers to differentiate between certain talents.
Talents such as Berserk: X for Druids and X Wrath for Paladins all impact the same spellid, so the talents have the same spellid, this means the regular way of identifying whether a player has the talent won't work. Instead we can look for the talentId (simply called id in CombatantInfo from WCL), and use that to determine whether a player has the given talent.

We could roll this out to be the standard way of doing hasTalent() but that goes a bit beyond the initial scope of this PR, unless it's really desired - if so I'll do that next.