fix broken lazy-loaded statistics
certain statistics like Power Word: Barrier use
LazyLoadStatisticBox in order to lazily load certain data
from WCL. when clicking on the box to load the statistic,
they are not currently updating with the appropriate value.

before the functional component rewrite of Results,
Results exposed a `forceUpdate` function via the legacy context
API. this `forceUpdate` function called a `forceUpdate` function
exposed by the React `Component` class. the `forceUpdate` function
caused a full-rerender of the Results component tree. once the
Results component was rewritten as a functional component,
access to this method was lost.

Analyzer instances are not React components, but they expose
multiple functions that can return React components. `statistic`
is one of these. React components returned by `statistic` are
combined in the `CombatLogParser#generateResults` function and
only update with data from their Analyzer instance if
`statistic` is invoked.

after tracing the render tree for the Analyzer `statistic`
function, I came to the conclusion that this below line
was no longer being invoked once the statistic was clicked,
causing the `statistic` function to subsequently not be invoked:
```
const results = isLoading ? null : props.parser.generateResults(adjustForDowntime);
```

as it is written, it is intended to be evaluated every time
the Results component is rendered. since we currently only render
the Results component when a full report/fight/player combination
is selected, this leads to Results only being rendered once (and
generateResults only being invoked once) - though clicking the
toggle that adjusts statistics for downtime can cause a re-render
of the Results component. this proved that a re-render of Results
would resolve the issue.

for functional components, re-renders typically happen when
props change or internal hook state changes occur. as mentioned
previously, functional components don't have a built-in
equivalent of `forceUpdate`. the React docs have a suggested
alternative using hooks (though they do caution against this
pattern):
```
const [ignored, forceUpdate] = useReducer(x => x + 1, 0);
```

I came to the conclusion that `forceUpdate` wasn't actually
what we want the LazyLoadStatisticBox component to cause.
what the LazyLoadStatisticBox really wants is to cause the
`CombatLogParser#generateResults` function to be evaluated.
as a result, my solution is to store the `ParseResults` from
the evaluation in state, evaluate it whenever the CombatLogParser
instance changes, and provide a React context that can be
consumed by LazyLoadStatisticBox to force a re-evaluation.
the evaluation whenever the CombatLogParser instance changes
allows the current desired behavior to be maintained while also
not invoking the "expensive" evaluation on every single prop
change.