remove Shadowlands spells (except food)
remove remaining Shadowlands spell references except
for food, which is used to make sure that people aren't
using Shadowlands food for their food buffs.

any commented code is due to lack of familiarity with
the spec and not wanting to delete code that may be of
use to future maintainers of those specs.