[Fire Mage] Supported for 10.0.5
Updated support for 10.0.5. This was supported a while back, but i never flipped it to supported since I hadnt had time to test. I have not seen any complaints for Fire in the Mage Discord, so I think it is safe to mark supported and update if any issues come up