handle RP "fights" for classic
there is a small chance that on rare occasions this could load
out-of-date combatantinfo. we could whitelist specific fights (Algalon,
Yogg-Saron, and Hodir all have this problem. Previously several fights
in SWP/BT did as well), but that requires maintaining a whitelist and
I'd rather not.

the consequences of using the wrong combatantinfo basically means
loading the wrong gear/talents (and thus potentially picking the wrong
spec)