set up GA4 + tag with fight/class/spec data
when i set up the ad script, I accidentally nuked our GA setup.

prior to this, we had GA's Universal Analytics set up but not set up
correctly---we weren't getting navigations from the router.

this fixes that mistake (and also moves us to GA4). the solution I went
with is *slightly* unconventional (often you see it recommended to do
this at the router level) because I want to capture some info about what
specs people are (trying to) look at, as well as whether they're looking
at dungeons or raids, etc.

this adds structured tags for class, spec, boss id, and dungeon vs raid.
some of the other tags (component_tag, for example) are mostly for
debugging this to make sure we're not getting tons of duplicate page
views