[SpellUsable] Adjust the last EndCooldown event
This change should alleviate some of the >100% cast efficiencies, since
the last EndCooldown event was scheduled at the cooldowns expected end
time (which could be past the fights duration) and that time from fight
end till EndCooldown was inadvertently added to the efficiency
calculation.

We now only look at the last EndCooldown event if it is within the
fight duration. If it is not, we take the remaining combat time as the
time spent on cooldown, whilst keeping the cooldown duration the full
duration.

We also remove the timeOffset from the cast efficiency calculation,
as it was made as a workaround for the above issue.