fix a race condition in useEventParser that comes up with phases a lot
when using `[value, setValue] = useState()`, calling `setValue` does not
update the actual contents of `value` until the next time the hook is
called.

the body of the main `useEffect` in this hook was defensive about that,
but if you selected a short phase (say 20s long) and then a long phase
(say 2m long), then the last `useEffect` would run, observe the old
value of `eventIndex` even after `setEventIndex(0)` along with the fact
that `eventIndex < normalizedEvents.length`. This would cause it to
begin processing the events in the new phase that occurred after (in
this example) 20s into the phase, then call `setEventIndex(currentIndex)`
with a non-zero value. The next time the hook was called, it would no
longer be 0 and the processing would not start over from the beginning.

the new setup is both more responsive (`requestIdleCallback` is better
about letting user input happen) and a bit saner (not so many
interleaving hooks that can cause re-renders)