Normalize enrage removal/applications to be refresh instead.
The timeline gets really messy because most abilities keeps replacing the buff. Also tracking haste changes becomes really messy because it is modified for every single attack. By normalizing them to refresh, we cleanup a lot of events and makes following operations simpler. For example, the normalizer to put the enrage events before the bloodthirsts will be simpler since we only have to reorder the Refresh event.