update talents for patch 10.2.5
this updates the names of several classes as well to match the new
talent names.

i started updating everything and uh, the degree to which that would be
invasively renaming a bunch of Windwalker / Mistweaver code made me pull
back. I renamed the main module and the WW / MW maintainers can handle
the rest.

I did confirm that both WW and MW still load. Warrior hasn't been
updated so no confirmation that it loads