add an ICON_RENAME map to handle mismatching PW:R names
the actual icon path in-game has a space in it. Raidbots data fills with
a `_`, while WCL & Wowhead both fill with a `-`.

I've reported the situation in the raidbots discord, will see what they
say.

This adds a general map that we can use to fix cases like this.

In 11.0, maybe we also use the WCL talent data files to make sure the
icon names match?