Prepare for Dragonflight season 4 and Cataclysm Classic (#6709)
* if the thumbnail response is a full URL, use it as-is

the new Blizzard API endpoints give us full URLs that are...fine.
they're headshots, we can use them.

open to discussion because they ARE different. its the difference
between the thumbnail image on a WCL character page and the full render
shown to the side.

* more compatibility updates for v2

- rankings needs className/specName now, so we set those for classic
  specs
- ranking structure changed, so handle that
- fix some spots i missed with thumbnail/avatar handling

i have NOT tested classic yet. that will likely need work

* switch to Vite over create-react-app

removes create-react-app and customize-cra setup in favor
of Vite setup, including sentry vite plugin.

* update dependencies

* fix TS2612 findings

* run prettier

* use babel to parse configs instead of trying to import them

our configs end up requiring a huge chunk of the codebase when eagerly
evaluated as part of an `import` statement, which has caused headaches
repeatedly.

so...i gave up and just used babel as a fancy macro system to parse out
the content that we needed to set up the e2e tests

* properly filter the tests like the old implementation

* don't try matching on spec name here

* fix undefined error when viewing w/ old server

* fix a few react "error" warnings that got changed to errors

* fix prettier error that snuck in

* re-add package that i accidentally deleted

* get e2e *actually* working

* 90% sure this works

* add the partition error to ignore list

Wrath got partitioned today

* bump sentry and TypeScript (#6679)

* bump sentry dependencies

* bump TypeScript to 5.4

* update dependencies in next branch (#6685)

* update dependencies

* fix prettier finding

* switch to react-virtualized-auto-sizer (#6689)

part of steps for WoWAnalyzer/WoWAnalyzer#6686

* fix broken talent references in Holy Priest changelog

---------

Co-authored-by: David Smith <emallson@emallson.net>