[Holy] 10.2.6 Changes (#6691)
* [Holy] Update Resonant Words

* [Holy] Update Healing Chorus

* [Holy] Update Prayerful Litany Values

* [Holy] Update Pontifex

* [Holy] Update Divine Word & Implement Mana Cost Savings

* [Holy] Fix Holy Word Mana Costs

* [Holy] Tidy up unused variables in Divine Word

* [Holy] Add EventLinks for Lightwell Renew and Salvation Renew
These are estimates at the overall healing done. To get an accurate number a custom Normalizer is needed

* [Holy] Add Lightwell
Using an estimate for the Renew healing by claiming the next 6s of renew healing after a Lightwell healing event. Due to the fact that the Lightwell renew is just extending any active renews this is not 100% accurate. Needs a custom Normalizer to handle renews.

* [Holy] Change Salvation to use Cast Links
This is a significantly better and more accurate means to estimate the Renew Healing done.
Effectively, any healing done by renew for 15 seconds after Salvation would have been done by Salvation, whether it ends up being applied by Benediction or other effects. So Assign it all to the Salvation CD.

* [Holy] Remove useless include

* [Holy] Benediction Improvement
As mentioned in previous commits this is still not a great way to do it.
Ideally everything goes back into Renew.tsx at the very least inside of a enum map based on the spells but an ideal involves the full rewrite as previously mentioned. Thus - Temporary quick bodge job to improve the accuracy significantly

* [Holy] Add Missing Resonant Words Spell

* [Holy] Add Contributors

* [Holy] Re-enable Changelog

* [Holy] set Flag isPartial

* [Holy] Fix Pontifex Comment