fix for the fix for source/targetIsFriendly (#6743)
* fix for the fix for source/targetIsFriendly

after initially confirming that my fix for this worked, I broke
it (oops).

while debugging this, I also discovered a subtle bug in the normalizer
that would improperly set some enemies as friendly because they would
not be in the `enemies` object yet since `getEntity` had not been
called. I rewrote bits of `FriendlyCompatNormalizer` to (a) make sure
that this does not happen, and (b) make sure that we're not doing a
huge amount of `find` calls for lookups of ids that are not enemies.

* late-introduced type error sigh

* fix test