Merge cata classic changes (#6751)
* refactoring to set up for Cataclysm Classic

i had to touch a bunch of places anyway to deal with the Wrath -> Cata
expansion change...so i just went ahead and moved some things around
while I was there.

- Classic specs now support Mastery. No coefficients have been set.
- Specs are now distinguished by their `branch`: `Retail` or `Classic`
- Moved many places that check an `Expansion` to instead check a `GameBranch`

The move from `Expansion` to `GameBranch` is not just tidying. It is anticipation of hitting MoP, at which point the line between "Retail" and "Classic" really starts to blur on the WCL side.

* improve hot refresh situation