fix FoF clipping detection
the issue is that the Channeling normalizer had started
running *after* the event link / event order normalizers.

normalizer order is arbitrary without explicit priorities or
dependencies. in this case, I used dependencies to force the
Channeling normalizer to run first.

i also fixed an issue where clipping with Chi Burst was not detected
because the `BeginCast` event type was not linked