add problems for marrowrend spam
this also adds some extra configuration options to the violation
problem list, and to the event order normalizer.

the change to the event order normalizer allows moving multiple events
via the same spec. this is useful for Bone Shield because a single
cast of Marrowrend may generate up to 4 events (apply + 2 stacks or
refresh + 3 stacks) and more than that during DRW