Expand Death Strike section of Blood DK guide (#6767)
* beginning revisions to the BDK guide

* initial work on suggestion box

these aren't wired up to anything, that is the next step

* make SuggestionBox keyboard-interactable

* some ui tweaks

* exclude defensive cds from the general cooldown list

* hook up the performance marks for runic power

* don't use !

* implement problem listing for RP waste

* rune cooldown info in guide

* add problems for marrowrend spam

this also adds some extra configuration options to the violation
problem list, and to the event order normalizer.

the change to the event order normalizer allows moving multiple events
via the same spec. this is useful for Bone Shield because a single
cast of Marrowrend may generate up to 4 events (apply + 2 stacks or
refresh + 3 stacks) and more than that during DRW

* improve explanations of mini-APL

* Druid - added detection of Season 4 set for Feral and Resto modules (#6764)

* [Fire Mage] Meteor Fix (#6766)

* Meteor Stuff

* changelog

* remove RP economy table from guide

this has been largely replaced by the DS section

* restore old usage section for now

* add changelog entry

---------

Co-authored-by: Kelton Finch <keltonfinch@gmail.com>
Co-authored-by: Sharrq <cjgrasty86@gmail.com>