temporary workaround for wcl talent breakage
the story is that WCL has changed its internal representation of talents
to no longer include the talent's ability id (since that can be
retrieved via the entry id). in the process, the character page was
changed to handle this and that had unexpected ramifications on this api
that need to be partially reverted.

this patch lets the page continue working (but without talent display)
until that happens