some initial setup for tww
did not add a new patch because we don't have a prepatch date yet. if
that breaks things i'll add one with a dummy beta timestamp to be fixed
later

raidbots dump has incorrect class ids and names right now. spec ids are
correct, but it'll be a bit of a pain to use those currently