temporarily disable the report selection tests til we have a properly supported spec
also fix a bug in the Arcane checklist that was causing a crash. this
bit of code is copy-pasted across a lot of checklists and i suspect is
the reason for that persistent error popping up in sentry so often
across so many logs (and being so hard to find).