rough draft of the foundation guide, using Cata BDK as a guinea pig
this adds a basic *Foundation Guide* that uses core analyzers to produce
some basic analysis with minimal additional code.

the goal is for you to be able to take this and plug it in to a
previously-unsupported spec to get something useful just by setting up
the core analyzers.

i ended up punting on the resource section for now because we don't have
any truly core (in the sense that they operate off of the Config or
Abilities) pieces for that just yet. However, I think that having a
shorthand to set up `ResourceTracker` / `ResourceGraph` will be a good
solution.