rough draft of the foundation guide, using Cata BDK as a guinea pig (#6832)
* rough draft of the foundation guide, using Cata BDK as a guinea pig

this adds a basic *Foundation Guide* that uses core analyzers to produce
some basic analysis with minimal additional code.

the goal is for you to be able to take this and plug it in to a
previously-unsupported spec to get something useful just by setting up
the core analyzers.

i ended up punting on the resource section for now because we don't have
any truly core (in the sense that they operate off of the Config or
Abilities) pieces for that just yet. However, I think that having a
shorthand to set up `ResourceTracker` / `ResourceGraph` will be a good
solution.

* text updates from jazminite

* switch caster template to foundation guide

* foundational support for Cata Destro Lock

* add "(copy definition)" helper to debug view

this can be used to quickly copy a spell definition from log data

* add Synapse Springs engineering item

* update healthstones and potion cooldowns for cata

* remove some low-rank spells from Cata Aff/Demo

* changelog

* update example report for destro lock

* more changelog

* update cooldowns

* review comments

* move classic healer template to foundation guide

* core setup for classic cata holy priest

* fix most spell errors from holy priest work

still need to handle Inner Focus

* re-add Inner Focus

the spell's semantics have changed for disc/shadow but that doesn't need
to be handled here

* implement mana section for healers

* fully remove low ranks from Priest/Warlock, text updates

this removal does NOT ensure that the remaining spell id is correct for
each spec.

this also fixes Divine Hymn being duplicated in Channeling---the spell
ID and buff ID are the same for Retail so I just left the retail one in
with a note about classic

* wrap foundation guide in "Core Skills" section

* little css tweak for immediate subsection